requirements.txt
streamlit==1.35.0
python-pptx==0.6.23
pandas==2.2.2
openpyxl==3.1.5
python-docx==1.1.2
pypdf==4.2.0
openai==1.40.6
anthropic==0.34.2
google-generativeai==0.7.2
python-dotenv==1.0.1
google-genai==1.29.0
pillow==10.4.0
